added API for table
